+++ src/main/java/com/telepathicgrunt/the_bumblezone/dimension/BzChunkGenerator.java
@@ -75,7 +75,7 @@ public class BzChunkGenerator extends ChunkGenerator {
     });
 
     private static final MobSpawnInfo.Spawners INITIAL_HONEY_SLIME_ENTRY = new MobSpawnInfo.Spawners(BzEntities.HONEY_SLIME.get(), 1, 1, 3);
-    private static final MobSpawnInfo.Spawners INITIAL_BEE_ENTRY = new MobSpawnInfo.Spawners(EntityType.BEE, 1, 1, 4);
+    private static final MobSpawnInfo.Spawners INITIAL_BEE_ENTRY = new MobSpawnInfo.Spawners(EntityType.BEE, 1, 2, 4);
     private static final BlockState CAVE_AIR = Blocks.CAVE_AIR.getDefaultState();
     protected final BlockState defaultBlock;
     protected final BlockState defaultFluid;
@@ -590,7 +590,7 @@ public class BzChunkGenerator extends ChunkGenerator {
         SharedSeedRandom sharedseedrandom = new SharedSeedRandom();
         sharedseedrandom.setDecorationSeed(region.getSeed(), xCord, zCord);
         while (sharedseedrandom.nextFloat() < biome.getMobSpawnInfo().getCreatureSpawnProbability() * 0.75f) {
-            //20% of time, spawn honey slime. Otherwise, spawn bees.
+            //25% of time, spawn honey slime. Otherwise, spawn bees.
             MobSpawnInfo.Spawners biome$spawnlistentry = sharedseedrandom.nextFloat() < 0.25f ? INITIAL_HONEY_SLIME_ENTRY : INITIAL_BEE_ENTRY;
 
             int startingX = xCord + sharedseedrandom.nextInt(16);